Enemy Spawn Animations +
Added in the enemy spawning animations and a condition/state so they have to finish them. With this i also added in the slot for their spawning sounds gotta actually get sounds for that and i also fixed the death hud so it should no longer show things like the reticle and points and what not.