Still Working On Charge
So i think i got the charge close to being done as  i want besides animations. But currently it charges gets the new lcoation but didnt charge to that new spot. So im currently implementing a stun period after a charge and then chaning that to having them going back to the chase. Gonna have to figure out the Explosion of the charges end location still.