Explosion update
Made the landing area explode in a larger area and made the particles show that off more. And currently i think it damages the player but not 100% sure and will be looking into this. Im also probably gonna make the explosion bigger as usually its a small radius around the enemy causing them so we shall see about that.